Add a *temporary* NO_WATCHDOG option to disable TimedExec.sh.
This is not the right fix. TimedExec.sh needs to be replaced by adding
a timeout option to the timeit tool. But, this has been sitting in my
copy of the test-suite for months and I would like to be able to tell
others to use the option. Disabling TimedExec.sh is the only way to get
meaningful run times for short benchmarks.

llvm-svn: 151237