Modify unix-small to be more appropriate as a performance test.
After reviewing the performance characteristics over time of the unix-smail 
benchmark with Daniel Dunbar, we determined that the majority of the run-time 
was spent inside of getpwent(), which is a library routine which gets 
information about users authorized to access the system the test runs on.  
This is not suitable for a performance benchmark, as the input may not be 
stable from one run to another.  The number of users may in /etc/passwd may 
change and/or the number of users authorized using some kind of network 
authorization mechanism may change as well.

The test is left in place as a correctness test.

llvm-svn: 154271