[tests] Modify TSVC static data layout to make benchmarks more stable.
As was previously written, the performance of these benchmarks can depend
significantly on the exact addresses assigned to the global data arrays at
runtime. Rewrite the tests to make the relative layout of each of the data
arrays fixed by storing them all in a single global structure. This makes the
behavior more predictable across platforms.

For example, previously on Mac OS X, two of the arrays would end up with
relative offsets at an exact multiple of 4K, which causes problems with some
Intel CPUs. This caused the performance of some benchmarks
(StatementReordering-flt in particular) to be dominated by a CPU specific hazard
instead of the actual benchmark code.

llvm-svn: 178968