[test-suite] Don't allow FP_TOLERANCE and HASH_PROGRAM_OUTPUT to occur toghether.
Summary:
LNT's benchmarks produce their output into files which are then compared in
various ways.

The benchmarks output numerical scores, which can then be compared by a custom
program in the test-suite tree which does a floating-point comparison with a
given tolerance (fpcmp).

The test-suite system allows allows users to do a hash of the program output
for a subsequently speedier comparison. This is used for benchmarks that
generate a significant amount of output (currently defined as >100k), The
problem is that some benchmarks that chose this feature also chose a
floating-point comparison of these hashes, which at best is nonsense, and at
worse causes an infinite loop.

The fpcmp program is not resilient when it's input is weird like this. For a
particular case of an observed mafft failure, it caused fpcmp to go into an
infinite loop.

Reviewers: kristof.beyls, ddunbar, rengolin

Subscribers: rengolin, llvm-commits

Differential Revision: http://reviews.llvm.org/D13283

llvm-svn: 248898