Recognize *-*-macosx* as Darwin
After r256038, we pass --target to configure from LNT.  LNT detects the
OS based on the target triple rather than uname for LLVM-based compilers
in lnt/testing/util/compilers.py.  This is macosx for us.

As a consequence, we didn't set TARGET_OS to Darwin in the testsuite
because the code only checked *-*-darwin until this change.  This led to
various testsuite failures.  E.g. SingleSource/UnitTests/2007-04-25-weak
was run which is normally disabled for Darwin.

rdar://problem/25565911

Approved by Chris in the radar.

llvm-svn: 265552