[test-suite] Add non-Linux fmemopen.c to CLAMR
For non-Linux systems, the upsteam sources have an implementation of fmemopen.
I'm adding it here (in the hope that it will make the Mac builders happy). In
the upstream sources, the CMake configuration builds this only on Darwin. It
would, however, also be needed on *BSD. Here, I've just put:

  #ifndef __linux__

around the source in the file.

llvm-svn: 312465