[test-suite] Set up .arcconfig to point to new Diffusion T repository
Summary:
We want to automatically copy the appropriate mailing list
for review requests to the test-suite repository.

For context, see the proposal and discussion here:

http://lists.llvm.org/pipermail/cfe-dev/2017-November/056032.html

Similar to D40179, I set up a new Diffusion repository with callsign
"T" for test-suite:

https://reviews.llvm.org/source/test-suite/

This explicitly updates test-suite's .arcconfig to point to the new
T repository in Diffusion, which will let us use Herald rule H270.

Reviewers: klimek, sammccall

Reviewed By: sammccall

Subscribers: llvm-commits, dlj, bkramer

Differential Revision: https://reviews.llvm.org/D40502

llvm-svn: 319667