cmake: Explicitely specify benchmark data
Explicitely declare what files are used as inputs for the benchmarks.
This changes the benchmarks to:

- Symlink (or copy) the related input files next to the binary into the
  build folder.
- Set the working directory of the benchmark to be the folder of the
  executable.
- Having the data next to the binary also reduces the amount of absolute
  paths in the test files.
- Contains some smaller cleanups for SPEC95/2000/2006 files while
  adapting to the new data copying style.

With this change in place you do not need the test-suite source
repository anymore to run the benchmarks. This is an important step
towards having a test-suite that can be cross-compiled on a host, then
copied onto a device and ran there without having a shared filesystem
like NFS setup.

Differential Revision: https://reviews.llvm.org/D50209

llvm-svn: 340582