[test-suite] Add Architecture Detection for RISC-V
Summary:
The LLVM test suite has its own way of detecting the system
architecture. This adds support in that file for detecting RISC-V.

This will eventually cause the ARCH variable to be populated.

We use ARCH="riscv64" to identify 64-bit RISC-V, and ARCH="riscv32" to identify 32-bit
RISC-V, so that attempting to detect "riscv" in the ARCH variable will match any version of RISC-V.

Reviewers: asb, luismarques

Reviewed By: luismarques

Subscribers: mgorny, simoncook, kito-cheng, shiva0217, rogfer01, rkruppe, PkmX, s.egerton, llvm-commits

Tags: #llvm

Differential Revision: https://reviews.llvm.org/D68698

llvm-svn: 374176