Rename test-suite test case to prevent NOEXE failures
In the test-suite, we have two files:
  SingleSource/UnitTests/Vector/Altivec/ld.reference_output
  SingleSource/UnitTests/Vector/Altivec/ld.c

That need to have their name changed to:
  SingleSource/UnitTests/Vector/Altivec/vec_ld.reference_output
  SingleSource/UnitTests/Vector/Altivec/vec_ld.c

Because they cause other test cases in the directory to not produce
executable once the ld test is built. The reason for the failure is that
the ./ld is used as the linker when building other tests.