Fix undefined behaviour in ldecod
The NUM_BLOCK_TYPES macro is defined to 10, but the arrays in
ctx_tables.h only have 8 elements in the relevant dimension, so this
was causing out of bounds array accesses, and segfaults if we were
(un)lucky. I don't know enough about H264 decoding to know if this is
the right fix, but it fixes the segfault currently affecting the Arm
LNT bot, and doesn't change the expected output.

Differential revision: https://reviews.llvm.org/D74770