[test-suite] Support GCC compilation of Microbenchmark MemFunctions.
GCC and Clang diverged in how __attribute__ is supported. GCC gives a compile-time error if there is an __attribute__ in the function-definition and needs to be in the forward declaration. Clang does not like __attribute__ in the forward declaration and requires it in the function-definition. This is the only benchmark in the test suite not compiling with GCC.

Reviewed By: courbet

Differential Revision: https://reviews.llvm.org/D93279