[simd_ops] Build tests with -fpic
Since https://reviews.llvm.org/rG2047c10c22b0, the dso_local specifier
must be included in IR when building non-PIC code. The bitcode files in
this test suite do not have this, so will only work when all of the code
is built with -fpic.

The alternative to this would be to re-generate the bitcode with the
dso_local specifier where needed, but I don't know how this code was
generated in the first place, and don't want to modify the bitcode files
in-place if they might be later re-generated.

Differential revision: https://reviews.llvm.org/D94085