[SPEC2017] Fix image_validator build when cross-compiling.
When cross-compiling and running benchmarks on a different device,
verification happens on-device, but currently the image validator
binaries are built for the host system.

This breaks verification, because the host binary cannot necessarily be
executed on-device (e.g. cross-compiling for ARM64 on X86 and executing
on a ARM64 device).

I tested this with cross-compiling to ARM64 and it works as expected.
When not cross-compiling, the behavior should not change.

Reviewed By: Meinersbur

Differential Revision: https://reviews.llvm.org/D95358