[SPEC2017] Use inputs from RUN directory.
At the moment, some benchmarks use inputs from the INPUT/SRC directory.
Those are not copied to the device, if one is specified. To ensure the
input is available on device, use it from the RUN_* directories and use
speccpu2017_prepare_rundir to copy them to the run directory. This will
automatically copy them to the specified device as well.

Reviewed By: Meinersbur

Differential Revision: https://reviews.llvm.org/D95360