[test-suite] Fix Unit Tests for Code Coverage
https://reviews.llvm.org/D88931 recently landed to add more tests to the
test suite to improve coverage of parts of clang. These tests caused
errors on some platforms, which was caught in post-commit review.

This change replaces the use of "%i" with the equivalent specifiers for
those types from inttypes.h, and updates the reference output to
correctly account for formatting unsigned arguments using unsigned
format specifiers.

Reviewed By: yroux

Differential Revision: https://reviews.llvm.org/D97686