[MicroBenchmarks] Add initial loop vectorization benchmarks.
This patch adds initial micro-benchmarks with interesting
loop-vectorization cases. To start with, it includes benchmarks using
libm math functions.

For each math function, there's a benchmark for the auto-vectorized
version and a version with vectorization disabled.

The auto-vec version of the benchmark also compares the results of the
auto-vectorized functions to the scalar versions.

Reviewed By: Meinersbur, lebedev.ri

Differential Revision: https://reviews.llvm.org/D101844