[SPEC2017] Copy reference files to run dir, use for compares.
When running the benchmarks on a different device, we need to copy the
reference files to the device and use the correct path for comparing on
the device.

We copy the output files (and compare files, if exist) to
${RUN_${_run_type}_DIR}/compares, which will be copied over to the
device, if one is configured. This is similar to how input files are
handled.

The validation commands have been updated to use the reference files
from the 'compare' directory in the run directory.

The patch also drops the WORKDIR argument. By dropping all uses of
WORKDIR, the inputs from the run directory are used, which should
work as expected when running the benchmarks locally and on a remote
device.

Reviewed By: Meinersbur

Differential Revision: https://reviews.llvm.org/D107061