Run llvm-size as part of test-suite build
This changes the test-suite build to run llvm-size on exectuables
immediately after they were built rather than only later when llvm-lit
runs the benchmarks.

This helps in situations where the benchmarks are built on 1 machine but
copied to a different machine to be run which may not have llvm-size
available at the same path or not at all.

Differential Revision: https://reviews.llvm.org/D108846