[test-suite] Change printf specifiers to capital
This patch changes some of the specifiers in the printf calls within:
```
MultiSource/Applications/sgefa/sgefa.test
MultiSource/Benchmarks/TSVC/CrossingThresholds-dbl/CrossingThresholds-dbl.test
MultiSource/Benchmarks/TSVC/CrossingThresholds-flt/CrossingThresholds-flt.test
```
to use the capital E/F specifiers. It also edits the reference files to account for the capital letter output.
This change was made because on AIX, inf is printed as INF, so the reference output and printf calls
are changed to print in capitals to avoid introducing a new reference output for AIX.

Differential Revision: https://reviews.llvm.org/D103068