[AIX][test-suite] Update the telecomm-FFT benchmark to use saved sin/cos values for some inputs.
This patch update the telecomm-FFT benchmark to use saved sin/cos values for
some inputs, to account for the differences of the AIX libm math library. When
compared to Linux libm, with the AIX version of the sin and cos functions
producing slightly less accurate results (with 1 ulp).

The AIX sin function produces different results when the input is (pi/4) or
(pi/16), whereas the cos function produces different results for the input
(pi/4). The Linux libm sin/cos results based off these inputs are saved
as three const-declared variables and substituted in when appropriate to ensure
the benchmark on AIX produces the same results.

Differential Revision: https://reviews.llvm.org/D107519