[test-suite][AIX] Disable SingleSource/UnitTests/2007-04-25-weak on AIX.
For input into the static linker, it is not possible to encode within an XCOFF object file
that the symbol is to be weakly linked. The AIX mechanism for symbols that may remain
unresolved after initial runtime linking requires the link step to be modified.
Additionally, the resolution of such deferred imports is less automatic on AIX.

As such, SingleSource/UnitTests/2007-04-25-weak test is currently not expected to
pass on AIX, so this patch disables building it and running on AIX.

Differential Revision: https://reviews.llvm.org/D115285