[test-suite] Add unit tests for vectorizer memory runtime checks.
This patch adds a first set of tests to check memory runtime checks
generated by the vectorizer.

The it runs scalar and vectorized versions of a loop requiring runtime
checks on the same inputs with pointers to the same buffer using various
offsets. It fails if they do not produce the same results.

The test functions are provided as lambdas, which are passed to a
driver function that generates the inputs and calls the lambdas with
pointers to overlapping buffers.

Reviewed By: Meinersbur

Differential Revision: https://reviews.llvm.org/D119121