[test-suite] Disable Fortran FM509 test when -std=legacy is not supported
This test requires -std=legacy which is not supported by flang.
Currently we check if the flag is supported and add if so,
but run the test anyway if not supported.

This has been failing on our bots for a long time but was not
noticed due to some issues reporting NOEXE failures.

https://lab.llvm.org/buildbot/#/builders/179/builds/3132

Executable Missing Tests (1):
  test-suite :: Fortran/UnitTests/fcvs21_f95/FM509.test

This fixes llvm_singlesource to allow you to pass a list
of source files (which used to work when it was a macro).

(this was done for multisource in d6f7103992e286a2ca9e68fbd82ca6dc570e6dc8)

If the flag is not supported we remove the test from the list
of source files. This means it's disabled for flang and
enabled for gfortran.

Reviewed By: awarzynski

Differential Revision: https://reviews.llvm.org/D121287