[Fortran] Re-enable FCVS FM905 & FM907
FM905 and FM907 have been disabled because they use list-directed
output, which gives the compiler some freedom regarding the exact format
and precision of the output. Our reference output had been obtained with
gfortran, and flang's output did not match.

This patch re-enables the tests, but switches the reference outputs to
match what flang produces. By default, we will check the output exactly.
However, the patch also introduces a new CMake variable,
FCVS_ALLOW_FLEXIBLE_OUTPUT, which can be set when running the
test-suite with other compilers (or even with other flags or environment
variables) and which will check the output more loosely. In this patch,
this only affects whether or not we ignore whitespace differences in the
output.

Note that this patch is not enough for the test-suite to pass with
gfortran, even when FCVS_ALLOW_FLEXIBLE_OUTPUT is enabled. Future
patches will try to fix that on a best-effort basis. Other fixes may be
required for other compilers or other compiler options.

Differential Revision: https://reviews.llvm.org/D131754