[AIX][test-suite]Workaround other benchmarks affected by the _ALL_SOURCE issue on AIX
On AIX, there are typedefs and struct definitions required by test-suite tests
that are only accessible when `_ALL_SOURCE` is set, but setting this macro
causes side effects such as typedef redefinition.

To fix them on AIX, we build test-suite with standard compatible mode
(e.g., XOPEN_SOURCE=700), and either add missing definitions to the tests or
adjust the tests to recognize that the extensions are disabled. This patch fixes
benchmarks that were not handled in D118622.

Differential Revision: https://reviews.llvm.org/D118660