[MicroBenchmarks,AArch64] Added correctness test & other performance tests for truncate or zero-extend vector operations
This patch adds a correctness test to check the outcome of vectorized truncate or zero-extend operations in a loop for different vector types.
This patch also adds performance tests for combined operations of truncate/zero-extend and addition.

The goal of this benchmark is to check the impact of AArch64 specific changes in
D133495, D120571, D135229 and D136722.

Reviewed By: fhahn

Differential Revision: https://reviews.llvm.org/D138059