[test-suite] Add strings.h header to ClamAV and define _XOPEN_SOURCE=600 on AIX for ClamAV/archie-client.
The ClamAV and archie-client tests are currently disabled on AIX, and we are now
in the process of re-enabling these tests and resolving the remaining issues.

As Clang now treats -Wimplicit-function-declaration and -Wint-conversion as
errors, ClamAV and archie-client are seen to exhibit error messages related to
calls to undeclared functions and incompatible integer to pointer conversions.
To resolve these issues on AIX, we simply need to include the appropriate
headers, which are <strings.h> and <netdb.h>. Although the <netdb.h> header is
already included, the gethostname() function is not visible as we previously
defined _XOPEN_SOURCE=700 on AIX in rT000c7b8d78b12b5cc70ad4e69a4cdb1bab23f74d.
Thus, for ClamAV and archie-client specifically on AIX, we define
_XOPEN_SOURCE=600 instead.

Differential Revision: https://reviews.llvm.org/D138610