[TSVC] Use fabs in s318 and s3113
In the s318 and s3113 kernels, the integer variant of the abs() function was
used to to determine the absolute value. The result is stored in a variable
'max' which has the float type, which makes me think the intention here was to
use fabs() variant, not the integer abs() variant.

Differential Revision: https://reviews.llvm.org/D138924