compare.py: increase --filter-short threshold, and accept optional argument
Option --filter-short now accepts an optional argument:

   --filter-short [FILTER_SHORT]
                        Filter benchmarks with execution times less than N seconds (default 1.0s)

It filters out applications with a runtime that are less than the provided
threshold value. If no value is provided, it defaults to 1.0 seconds, which
used to be 0.6 seconds.

The rationale is that the LLVM test-suite and performance results can be quite
noisy and this is one mitigation for that. The swings for sub-second benchmarks
can be quite big, and this hopefully allows to do less runs to get stable
numbers.

Differential Revision: https://reviews.llvm.org/D144101