[gcc-c-torture] Note undefined behaviour in frame-address.c
In 6ae5ef2a7b64a57278eadaa8ab6c35031331e045 I disabled this test
due to clang now tail calling one of the functions.

In addition, the callee compares pointers to different objects
which is undefined behaviour. I've added this to the comment
in case we look at enabling it again later.

Clang and GCC for now do the same thing with it, but it's
not guaranteed.