[test-suite] Avoid segfault in CLAMR test on non-Linux systems
On FreeBSD, the CLAMR test immediately segfaults on startup, when it
parses /proc/pid/status in the `memstats_memused()` function. This is
because the parser is not very robust, and the FreeBSD /proc format is
different than Linux.

I expect this code to crash on anything else than Linux, but for our
test purposes the memory statistics printed are not important, since
nobody ever looks at the produced output.

Since there is already a `MEMORY` define controlling the printing of
memory statistics, only call the `memstats_memused()` function if it is
explicitly enabled (which it is not, by default).

Reviewed By: brad

Differential Revision: https://reviews.llvm.org/D146387