[Fortran] Comment out tests with missing semantic checks (#55)
These tests are looking for a semantic error that flang is not emitting
but they used to pass because a TODO for procedure pointers was emitted
in lowering. Lowering implemented procedure pointers, so these test
now fail. I opened an issue against semantics, but I want to comment
out these tests so that we can still land the lowering implementation
of procedure pointers that is correct.

See https://github.com/llvm/llvm-project/issues/73215